+++ CHANGELOG.md
@@ -5,6 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## [Unreleased]
 
+## [0.2.0] - TBD
+### Breaking
+- `XvizPanel` is renamed to `XVIZPanel`.
+  + `id` prop is renamed to `name` to align with the Declarative UI spec 
+- `VideoPanel` is removed. Use declarative UI with `XVIZPanel` instead.
+
 ## [0.1.0] - 2018-10-24
 ### Changes
 - Update for XVIZ 2.0 spec compliance and @xviz 0.1.0